refactor: Move config and constants imports to core module
